- update portable reader events and add faketag - update logic when portable reader exe does not exist

+++ src/lib/readers/portable-reader.js
@@ -1,10 +1,13 @@
 const { spawn } = require('child_process');
 const EventEmitter = require('events');
+const fs = require('fs');
 
 const pREADER_MSG = {
     START_LISTEN: 'start_listen\r\n',
     CONTINUE_LISEN: 'continue_listen\r\n'
 };
+
+const EXE_FILE_PATH = '/bin/portablereader.exe';
 
 class PortableReader extends EventEmitter {
     constructor() {
@@ -19,7 +22,7 @@ class PortableReader extends EventEmitter {
         }).catch(err => {
             console.log(err);
         });
-        
+
     }
 
     continue() {
@@ -27,39 +30,53 @@ class PortableReader extends EventEmitter {
     }
 
     kill() {
+        if (!this.process) {
+            return;
+        }
+
         this.process.kill();
         console.log('Portable reader process was killed');
     }
 
+    fakeTag() {
+        this.emit('tag', 'FAKE_TAG_UID:123456789');
+    }
+
     _open() {
+        this.emit('connectingStart');
+        const msg = `${EXE_FILE_PATH} not found`;
+        if (!fs.existsSync(EXE_FILE_PATH)) {
+            this.emit('connectedFailed', msg);
+            return Promise.reject(msg);
+        }
+
         return new Promise((resolve, reject) => {
-            this.emit('connectingStart');
             const delay = '200';
-            this.process = spawn(process.cwd() + '/bin/portablereader.exe', [delay]);
+            this.process = spawn(process.cwd() + EXE_FILE_PATH, [delay]);
             this.process.stdout.on('data', data => {
                 const [status, message] = data.toString().trim().split(':');
-    
+
                 if (status === 'error') {
                     this.emit('connectedFailed', message);
                     reject('Connected to portable reader failed, message: ', message);
                     return;
                 }
-    
+
                 if (status === 'ok' && message === 'connected') {
                     this.isConnected = true;
                     this.emit('connected');
                     resolve();
-    
+
                     return;
                 }
-    
+
                 if (status === 'tag') {
                     this.emit('tag', message);
-    
+
                     return;
                 }
             });
-    
+
             //TODO CATCH ERROR IF READER DISCONNECT
             this.process.on('close', (code, signal) => {
                 console.log('Portable reader process was closed', code, signal);
